+++ demo/demo.js
@@ -1,9 +1,9 @@
 // Character spritemap is an edited version of the original made by Buch: https://opengameart.org/users/buch
 
-import Engine from "../src/core/Engine";
-import { vec } from "../src/math/Vector";
-import Actor from "../src/objects/Actor";
-import { TextureLoader } from "../src/util/TextureLoader";
+import Engine from "../src/core/engine";
+import { vec } from "../src/Math/vector";
+import Actor from "../src/objects/actor";
+import { TextureLoader } from "../src/util/texture_loader";
 
 import animationSpritemap from "./animationSpritemap.png";
 import characterSpritemap from "./characterSpritemap.png";
