+++ src/core/event_subscriber.ts
