+++ src/util/texture_loader.ts
